remove old useFieldTree
