--- app/src/composables/use-field-tree-advanced/index.ts
@@ -1,4 +1,0 @@
-import useFieldTreeAdvanced from './use-field-tree-advanced';
-
-export default useFieldTreeAdvanced;
-export { useFieldTreeAdvanced };
