--- app/src/composables/use-field-tree/readme.md
@@ -1,21 +1,0 @@
-# `useFieldTree`
-
-Generate out a field tree based on the given collection.
-
-```ts
-type FieldTree = {
-	field: string;
-	name: string;
-	children?: FieldTree[];
-};
-
-useFieldTree(collection: Ref<string>): { tree: FieldTree }
-```
-
-## Usage
-
-```ts
-const collection = ref('articles');
-
-const { tree } = useFieldTree(collection);
-```
